Changes Date Format for the JTable Parking and changing dates.
DaoParkingImpl: import SimpleDateFormat to format the Date we need to
display in the JTable. format(dd.MM.yyyy).

DaoParkingImpl: important changes by displayListParking(inside process
Method by the SwingWorker dates will be formated with the help of
SimpleDateFormat before to displaying in the JTable). By the same Method
we have change the code to setCellRedere getting the correct
ABREISEDATUM and ANZAHLDESTAGES value.

DaoParkingImpl: the same Date format for the Method reloadParkingData.

DataBaseGUIController: import MouseEvent and MouseListener.

DataBaseGUIController: parkingTAble and fundsachenTAble add MousListener
and FocusListener.

DataBAseGUIController: Very important instructions by the keyPressed
Method. In case parkingTable we evaluate if Column 4 we get the value of
this column to set String value. Store the Column value in one variable
so in case something go wrong with editing we can get back the old value
easily.

DataBAseGUIController: tableChanged Method. We pass a Column argument to
the updateParking method call.

DataBAseGUIController mousePressed Method: in case double click we also
store the value of the column 4 in case something go wrong by editing.

LogicModel: import also the SimpleDateFormat class.

LogicModel: new Method. checkDataFormatJTableLostFocus to evaluate
before saving data if the date format is correct.

LogicModel: new Method getSimpleDateFormat, getLocalDAteToBeModified,
setLocalDateToBeModified, getDataAsStringToBeModified,
setDataAsStringToBeModified.

LogicModelParking: Big changes.

LogicModelParking: Create some new Objects to display error message in
case Date format is not correct.

LogicModelParking: We removed the Empty constructor.

LogicModelParking: new Parameter by the updateParking Method.

LogicModelParking: by the updateParking Method will be evaluated the
DateFormat and we set the LocalDates values. We also call a method here
checkChronologyOfDates to check if the modified Arrival Date is later
than departure.

LogicModelParking: Override checkDAteFormatBeforeSaveData Method.

LogicModelParking checkChronologyOfDates has all the instructions to
first store the values and then to call the DaoParkingImpl Object
passing the ParkingReservation Object with all the values stored to be
modified.

LogicModelParking: I am going to continue working in the
checkChronologyOfDates in case Arrival is Later then departure so we can
give the option to the user to modified the departure date.

!! I will continue the comments and some code modifications. Have to go
to the official Job. LOL.