Changes by modifying Parking and Lost and found.
DataBaseGUI:

- New Object JComboBox<String> to indicate if the car key will be kept
in House or not.

- In case PARKING the object choices initialize with some new values.
The JComboBox initialize with the choices object(Array as String).

- parkingTable get the Column 10 and setCellEditor(the
schluesselInHausJComboBox).

DataBaseGUIController:

- addItemListener to schluesselInHausJComboBox and kisteNummerJComboBox.

- We removed the Throws DaoException when we call for the update JTable.

- itemStateChanged we use conditional if.

LogicModelFundSachen:

- new instruction conditional if in case selectedRow !=-1 to evaluate if
a row was selected or not. Then we do not use any more in this Method
the catch block ArrayIndexOutOfBoundsException.

- by the switch block when we evaluate kisteN for each case we give
value to the variable kisteName.

LogicModelParking:

- We declare the variables for the parking reservation in the class it
self.

- By the updateParking Method will not used any more the try catch block
ArrayIndexOutOfBoundException. we call direct to evaluate if a row was
selected or not.

- by the updateParking Method also becomes new instructions inside the
switch block(selectedColumn).

- new Method updateAllParkingChanges used in case we do not need to
calculate any Total Days. That means Arrival Date and Departure Date was
not modified thats why this method could be called.