Parking JTable add JComboBox for the Cell Editor.
DataBaseGUI:

add new JComboBox element and String Array to use as cell editor by the
parkingJTable to modify the Sales Channel.

Initialize the choicesBuchugskanal array for selecting a new Sales
Channel. The buchungskanalJComboBox receives the values from the Array
and we set the CellEditor with the JComboBox Object.

DataBaseGUIController:

We add the ItemListener for the buchungskanalJComboBox.