Getting all the User for the User Manager GUI so we can edit, add...
PropertiesReader: New Method readPropertiesData to return our
configuration file properties so we can access from the User Manager to
the different users to edit, modify password or just delete user.

UserManager: We Use one JList to display all the user got from the
properties. The configuration file has all the user. We iterate with the
configuration file and if one user is found will be added to the JList
so we can work with this information.

We add the JList to the Panels.