New instructions to save the New User in the configuration file
LogicModelNewUser:

- import DataEncryption and PropertiesWriter to encrypt and to write the
new properties in the configuration file.

- New instructions in the saveNewUser Method to write the new user.


LogicModelUserManager:

- New Method addNewuserToTheList. This method should add the new User to
the GUI List where you find the actual user displayed.

UserManager: new Method addNewUserToList that calls the model to add the
element.


Configuration file was modified by this first test but we have some
bugs. after adding the new user in the list will be displayed the user
but after reload the application it will be displayed empty values. I
have to check the code again and repair in the next commits.