Adding new Elements to the NewUser GUI.
JComboBox to select the User Rights.

I create also by the NewUser Class String array with the choices we can
select for the New User: Manager(Administrator) or Mitarbeiter(Staff).

I am going to create for the next commits one more element for the
Shortcurt name and also will be fixed the bugs by saving the Data in the
configuration File.