New variables for the user rights and new method to check the rights.
LogicModelNewUser:

- New String variables userRights to store the user rights and
newUserRightsCorrect to control or check if the user has selected user
rights for the new user.

- By the Method saveNewuser it will be also checked if
newuserRightsCorrect is true calling the new Method
validateNewUserRights.

- I have commented for the moment the instructions to write the
properties in the configuration file until i write the other
instructions for checking also the abbreviation name for the new user.

- new Method: validateNewUserRights. Return newUserRightsCorrect value.

NewUserController Class:

new instructions or arguments by calling the saveNewUser Method by the
actionPerformed Method and by the keyPressed Method.