New instructions to save the new User in the configuration File fixed.
LogicModelNewUser:

- new Variables to store the user privilege or rights, the user
abbreviation name.

- By the saveNewUser add a switch block to evaluate this.userRights. In
case Manager will be replaced for ADMIN and Mitarbeiter replaced for
STAFF so when we logging the Applications display more JButtons or Not
depending if is one Administrator or only Personal Staff.

- Bugs fixed to store the properties name and value.

- Finally we write the new user properties in the configurations file.

LoginController:

checkPrivilegeUser Method evaluate Case Staff. And
BigDataAHBStartFrameController do the same evaluation.

Removed unused variables.