New Class to edit existing User in the configuration File:
- EditUserController
- EditUserGUI
- LogicModelEditUser

New instructions:

LogicModelUserManager: New Method editUser(I have not written any
instructions for editing users yet. Just one JOptionPane message).

UserManager: JList<String> list object visibility was changed to public
so we can access to the JList to get the selected item to edit.

UserManagerController:

- New ImageIcon element to display with the error message.
- Add Action Listener to the editUserJButton by userManager GUI.

- by the actionPerformed Method in case editUserJButton pressed we call
the editUser Method belonging to the logicModelUserManager instance. It
will be evaluated if item was selected or not to display an error
message or not.