New instructions to modify the user password and more...
BigDataAHBStartFrameController:

- import ModifyPasswordUserGUI;

- new String object passwordPropertieName for the password property
name;

- new Instructions by pressing btn_kontoVerwalten button(STAFF USER). In
case of ADMIN or STAFF it will execute different instructions.

LogicModelModifyPasswordUser:

- new String Object errorMessage to display error message by checking
the passwords;

- New constructor for the Class. This new constructor will be called
from BigDataAHBStartFrameController class. This constructor do not have
the parameter EditUserGUI because the ModifyPasswordUserGUI class will
be called in this case from the main Main GUI;

- new Instructions by the modifyPasswordUser Method called(The Second
one called by pressing the btn_kontoVerwalten(Manage own account).

- Some new Methods to check if all the passwords(old, new and
confirmation password) are correct or not.

ModifyPasswordUserGUI:

- new constructor used when this class will be called from the main GUI.
Also it receives BigDataAirportHotelBaselStartFrame Parameter.

PropertiesReader:

- setPassword to the userAHB instance. need it to access from main GUI
when we want to modify password.