working on changes for the correction of modified dates within the JTable. In case of incorrect chronology or incorrect date format. This Job is not finished jet.
DataBaseGUIController:

- New String Array date to store the 2 Dates we are able to modify. Is
used to get this value and in case something got wrong we can reset this
value back to the JTable. The value will get for the corresponding
columns depending which JTable is displayed.
The value for the array dates will be applied inside the Method
keyPressed and mousedPressed double Click by this DataBsaseGUIController
class.

DateChronologyCorrectionController:

- by the Method windowClosing will be called the resetDates method. That
means the user close the GUI(JDialog) without saving any modification.

LogicModel:

- new Array String dateAsStringToBeModified used to store the  old value
of the dates we are trying to modify.

- The Method getDateAsStringToBeModified return now one String Array
with the two Dates value stored inside.

- The Method setDateAsStringToBeModified call to store one Array by the
variable dateAsStringToBeModified.

- New Method: displayRequestDateCorrection.

Extra info for the LogicModel Class: I am modifying the instructions
displayRequestDateCorrection to finally delete the Method
displayRequestLaterDateCorrection. The new Method
displayRequestDateCorrection will be valid for other classes with JTable
access and working with two different Dates and Chronology.

LogicModelDateChronologyCorrection:

- New variables to store the old values of the two Dates as String
format dd.mm.yyyy.

- new Method resetDates to set back the old values of the Dates as
String dd.mm.yyyy depending which column was trying to modify the user.

LogicModelParking:

- Call now the method displayRequestDateCorrection inside the Method
checkChronologyOfDates. by the message we use in the String <html> Tags
so we can adapt better the message with new line etc.