New modifications in case user type wrong Date Format not dd.mm.yyyy
LogicModelParking:

 - New modification by the checkDateFormatBeforeSaveDate. Now for
setting the value we have to indicate the value of the Array String
position. this getDateAsStringToBeModifed Method return one String Array
with the to value of Dates as String. Format dd.mm.yyyy to be replace
back to the JTable in case the user entered a wrong date format as
String.