Some new checks and instructions to check the Chronology of Dates
Some new instructions to check if the Date format is correct before
applying the changes to the JTable(Parking) and to save the changes in
the DataBase.

DateChronologyCorrection:

- JTextField dateJTextField changed visibility to public so we can use
the keyPressed Method by the Controller Class.

DateChronologyCorrectionController:

- Add KeyListener to dateJTextField;

- New instructions by actionPerformed Method and keyPressed Method to
check if the Departure date that user is trying to modify is correct
format or not and if yes we call the setDepartureDate Method by
logicDateChronologyCorrection Class.

LogicModel Class:

deleted:

- variables or Objects TableModel model, dialogDateChange(JDialgo),
saveJButton, abbrechenJButon, Object Array
optionsRequestLaterDateCorrection, ImageIcon(infoImg).

- by the Method deleteRowDataBase we call the selectedRow variable using
static form.

- Method displayRequestDateCorrection removed some parameters(TableModel
model and int columnToBeModified.

- by the Method displayRequestDateCorrection the created
LogicModelDateChronologyCorrection instance will be created passing a
new argument dateChronologyCorrection so we can close this
GUI(dateChronologyCorrection) when the Job is done.

LogicModel VERY IMPORTANT: It was deleted the
displayRequestLaterDateCorrection. We use now official the
displayRequestDateCorrection because is more easy to access and remember
this method when we work with two dates to be checked from another
JTable.

LogicModelDateChronologyCorrection:

- extends now from LogicModel so we can access some Methods necessary
from LogicModel Class.

- New Object DateChronologyCorrection and new parameter
DateChronologyCorrection by the constructor.

- New Method setDepartureDate to set the value in the JTable when
everything is correct.

LogicModelParking:

by checkChronologyOfDates Method the call to the
displayRequestDateCorrection will be called with 2 arguments less.