New instructions to set also the (.) by abbreviation name.
In Some Logic Class new small modifications to save the abbreviation
name Property separate with a (.).

In all the modified class that request Abbreviation
Name(db.abkuerzungma.user it was applied after the user one (.) to the
abkuerzungPropertieName value so we save now with a . + user in the
configuration File.

This Property will be loaded also using the .