New instructions to display dates interval(Between Dates) DataBase Lost And Found.
Added new instructions to request SQL to display interval dates between
one date to others. Just to display by Lost and Found the Current month.

DAOFundsachen:

- displayListFundsachen Method has a now String parameter monthToShow.

DaoFundsachenImpl:

- new String variable monthToShow to store the value with the Month we
want to be displayed.

- modification by the checkTableFundsachen Method where we do not call
the displayListFundsachen Method any more, instead of that we call a new
Method(displayListfundsachenByMonth.

- by the Method addNewFundsachen will be called also the
displayListfundsachenByMonth method.

- By the displayListFundsachen Method becomes an String parameter
monthToShow. We evaluate with one switch block the value of monthToShow
and depending of the value the SQL instructions change.

- new Method displayListfundsachenByMonth.

-  searchByDateFundsachen call also the displayListfundsachenByMonth. So
we have optimized the code by this Method.

- Method reloadFundsachenData was also optimized and call the
displayListfundsachenByMonth Method where it will be displayed the
results with the current month entries in Database.