Initialize displayMothLostAndFoundJComboBox by the DatabaseGUI Class and set all values to display in the GUI and more..
The displayMothLostAndFoundJComboBox will be initialized in the
DataBaseGUI so we can insert this element in one
JPanel(containerMonthSelected) and this inside the loginPanel.

DatabaseGUIController:

- displayMothLostAndFoundJComboBox add ItemListener.

- New instruction by itemStateChanged for
displayMothLostAndFoundJComboBox so we call the method to display the
selected moth only if item was selected.

LogicModelFundSachen:

- new String variable monthSelected to store the Month we want to
display.

- new Method displayMonthSelected.

SelectCustomJComboBox:

- new Color for the setSelectionBackGround(We can in the future of
course send the color what we have when we initialize the Object).

TopPanelFundSchen: removed displayMothLostAndFoundJComboBox element.