Deleted the + LocalDateTim now.getYear option for the name of DataBase.
I decided to delete or not add any more the current Year to the
DataBase. So it will be kept only one DataBase and not the idea to crate
one DataBase per Year.

What i want to do in Future is. Some tables do not keep Data forever. So
when some Data depending which Table is older than 3 Years we delete the
oldest Data that we do not need anymore in some cases.

Tables like Fitness Subscription We can keep till 3 Years. If the
Fitness User do not come in 3 Years we delete the profile with all the
information to save place in local Disk.

DAOFitnessImpl: removed the System.out.println instruction.

LogicModel: removed the Year from DataBase name.