Instructions to delete Entries from the Lost And Found.
DataBaseGUI:

- import JMenuItem and JPopupMenu;

- Initialize the JPopupMenu and MenuItem.

DataBaseGUIController:

- implements PopupMenuListener;

- New variable String privilege to store the value privilege of the
user;

- DataEncryption to decrypt the value from the configuration File;

- dataBaseGUI.deleteItem add ActionListener;

- by the Method actionPerformed: case deleteItem: only if
fundsachenTable is present it will be evaluated if any row was selected.
If no row was selected we invoke a Thread with error message. If row was
selected we call logicModelFundSachen.deleteRowDataBase(method belonging
to the LogicModel Mother Class).

- by the windowOpened Method: we get the value of privilege. We also
evaluate if the fundsachenTable is present(Lost And Found JTable). If
this table is not null we setComponentPopupMenu. (VERY IMPORTANT because
this component will be set only also if privilege is equals ADMIN, so
that only ADMIN can delete any row from the DataBase.

LogicModel:

- changed String tableName to static;

- new ImageIcon preventionImage to display the icon in the alert
message;

- new instructions by the deleteRowDataBase Method so we request the
user if he is sure he wants to delete the selected entry.